requirements/production.txt
amqp
anyjson
Cython
Django==1.7.9
django-adminactions
django-admin-keyboard-shortcuts
django-bootstrap-pagination
django-celery
django-compressor==1.4
django-crispy-forms
django-extensions
django-filter==0.9.2
django-mail-queue
django-mptt
django-reversion==1.8.5
django-selectable
django-widget-tweaks
-e git+https://github.com/coagulant/django-tastypie.git@feature/django-1.7#egg=django-tastypie
mimeparse
nodeenv
psycopg2
python-dateutil
python-digest
pytz
rapidpro-for-rapidsms
rapidsms==0.18.0
raven
requests
tablib
Unidecode
